+++ lib/di.dart
@@ -5,12 +5,12 @@ import 'package:sovarpc/blocs/debug_rpc_bloc.dart';
 import 'package:sovarpc/blocs/noso_network_bloc.dart';
 import 'package:sovarpc/blocs/rpc_bloc.dart';
 import 'package:sovarpc/repository/file_repository.dart';
-import 'package:sovarpc/repository/repositories_rpc.dart';
-import 'package:sovarpc/services/file_service_rpc.dart';
-import 'package:sovarpc/services/settings_yaml.dart';
-import 'package:sovarpc/services/noso_network_service.dart';
 import 'package:sovarpc/repository/local_repository.dart';
 import 'package:sovarpc/repository/noso_network_repository.dart';
+import 'package:sovarpc/repository/repositories_rpc.dart';
+import 'package:sovarpc/services/file_service_rpc.dart';
+import 'package:sovarpc/services/noso_network_service.dart';
+import 'package:sovarpc/services/settings_yaml.dart';
 
 import 'database/database.dart';
 import 'models/log_level.dart';
@@ -26,7 +26,7 @@ Future<void> setupDiRPC(AppPath pathApp,
   locatorRpc.registerLazySingleton<LogLevel>(
       () => logLevel ?? LogLevel(level: "Debug"));
   locatorRpc
-      .registerLazySingleton<SettingsYamlHandler>(() => SettingsYamlHandler(pathApp));
+      .registerLazySingleton<SettingsYamlHandler>(() => SettingsYamlHandler());
 
   /// shared & drift(sql)
   locatorRpc.registerLazySingleton<LocalDatabase>(() => LocalDatabase(pathApp));